wip typecheck point
