+++ parser-typechecker/unison-parser-typechecker.cabal
@@ -45,7 +45,6 @@ library
       Unison.Codebase.Editor.HandleCommand
       Unison.Codebase.Editor.HandleInput
       Unison.Codebase.Editor.Input
-      Unison.Codebase.Editor.InputOutput
       Unison.Codebase.Editor.Output
       Unison.Codebase.Editor.Output.BranchDiff
       Unison.Codebase.Editor.Output.DumpNamespace
